+++ engine/flink/api/src/main/scala/pl/touk/nussknacker/engine/flink/api/NkGlobalParameters.scala
@@ -81,4 +81,3 @@ object NkGlobalParameters {
   }
 
 }
-